+++ js_modules/dagit/packages/core/src/pipelines/SidebarOpDefinition.tsx
@@ -4,6 +4,7 @@ import * as React from 'react';
 import {Link} from 'react-router-dom';
 import styled from 'styled-components/macro';
 
+import {useFeatureFlags} from '../app/Flags';
 import {breakOnUnderscores} from '../app/Util';
 import {displayNameForAssetKey, isHiddenAssetGroupJob} from '../asset-graph/Utils';
 import {assetDetailsPathForKey} from '../assets/assetDetailsPathForKey';
@@ -12,6 +13,7 @@ import {pluginForMetadata} from '../plugins';
 import {CONFIG_TYPE_SCHEMA_FRAGMENT} from '../typeexplorer/ConfigTypeSchema';
 import {DAGSTER_TYPE_WITH_TOOLTIP_FRAGMENT, TypeWithTooltip} from '../typeexplorer/TypeWithTooltip';
 import {RepoAddress} from '../workspace/types';
+import {workspacePathFromAddress} from '../workspace/workspacePath';
 
 import {Description} from './Description';
 import {
@@ -45,6 +47,9 @@ const DEFAULT_INVOCATIONS_SHOWN = 20;
 
 export const SidebarOpDefinition: React.FC<SidebarOpDefinitionProps> = (props) => {
   const {definition, getInvocations, showingSubgraph, onClickInvocation, repoAddress} = props;
+
+  const {flagSidebarResources} = useFeatureFlags();
+
   const Plugin = pluginForMetadata(definition.metadata);
   const isComposite = definition.__typename === 'CompositeSolidDefinition';
   const configField = definition.__typename === 'SolidDefinition' ? definition.configField : null;
@@ -113,7 +118,18 @@ export const SidebarOpDefinition: React.FC<SidebarOpDefinitionProps> = (props) =
             {[...requiredResources].sort().map((requirement) => (
               <ResourceContainer key={requirement.resourceKey}>
                 <Icon name="resource" color={Colors.Gray700} />
-                <ResourceHeader>{requirement.resourceKey}</ResourceHeader>
+                {flagSidebarResources && repoAddress ? (
+                  <Link
+                    to={workspacePathFromAddress(
+                      repoAddress,
+                      `/resources/${requirement.resourceKey}`,
+                    )}
+                  >
+                    <ResourceHeader>{requirement.resourceKey}</ResourceHeader>
+                  </Link>
+                ) : (
+                  <ResourceHeader>{requirement.resourceKey}</ResourceHeader>
+                )}
               </ResourceContainer>
             ))}
           </Box>