+++ js_modules/dagit/packages/core/src/ui/Filters/useStaticSetFilter.tsx
@@ -217,7 +217,7 @@ function SetFilterLabel(props: SetFilterLabelProps) {
       margin={{left: 4}}
       style={{maxWidth: '500px'}}
     >
-      <Checkbox checked={isActive} size="small" />
+      <Checkbox checked={isActive} size="small" readOnly />
       <Box
         flex={{direction: 'row', alignItems: 'center', grow: 1, shrink: 1}}
         style={{overflow: 'hidden'}}