+++ python_modules/dagster/dagster/_config/structured_config/__init__.py
@@ -67,7 +67,7 @@ from dagster._core.storage.io_manager import IOManager, IOManagerDefinition
 from .typing_utils import BaseResourceMeta, LateBoundTypesForResourceTypeChecking
 from .utils import safe_is_subclass
 
-Self = TypeVar("Self", bound="ConfigurableResource")
+Self = TypeVar("Self", bound="ConfigurableResourceFactory")
 
 
 class MakeConfigCacheable(BaseModel):
@@ -378,7 +378,7 @@ class IOManagerWithKeyMapping(ResourceWithKeyMapping, IOManagerDefinition):
 def attach_resource_id_to_key_mapping(
     resource_def: Any, resource_id_to_key_mapping: Dict[ResourceId, str]
 ) -> Any:
-    if isinstance(resource_def, (ConfigurableResource, PartialResource)):
+    if isinstance(resource_def, (ConfigurableResourceFactory, PartialResource)):
         return (
             IOManagerWithKeyMapping(resource_def, resource_id_to_key_mapping)
             if isinstance(resource_def, IOManagerDefinition)
@@ -388,7 +388,7 @@ def attach_resource_id_to_key_mapping(
 
 
 @experimental
-class ConfigurableResource(
+class ConfigurableResourceFactory(
     Generic[TResValue],
     ResourceDefinition,
     Config,
@@ -399,16 +399,16 @@ class ConfigurableResource(
     """Base class for Dagster resources that utilize structured config.
 
     This class is a subclass of both :py:class:`ResourceDefinition` and :py:class:`Config`, and
-    provides a default implementation of the resource_fn that returns the resource itself.
+    must implement ``create_resource``, which defines the resource value to pass to user code.
 
     Example:
     .. code-block:: python
 
-        class WriterResource(ConfigurableResource):
-            prefix: str
+        class DatabseResource(ConfigurableResourceFactory[Database]):
+            connection_uri: str
 
-            def output(self, text: str) -> None:
-                print(f"{self.prefix}{text}")
+            def create_resource(self, _init_context) -> Database:
+                return Database(self.connection_uri)
 
     """
 
@@ -458,7 +458,9 @@ class ConfigurableResource(
         """
         return PartialResource(cls, data=kwargs)
 
-    def _with_updated_values(self, values: Mapping[str, Any]) -> "ConfigurableResource[TResValue]":
+    def _with_updated_values(
+        self, values: Mapping[str, Any]
+    ) -> "ConfigurableResourceFactory[TResValue]":
         """Returns a new instance of the resource with the given values.
         Used when initializing a resource at runtime.
         """
@@ -467,7 +469,7 @@ class ConfigurableResource(
         # of this class. We can therefore safely pass in the values as kwargs.
         return self.__class__(**{**self._as_config_dict(), **values})
 
-    def _resolve_and_update_env_vars(self) -> "ConfigurableResource[TResValue]":
+    def _resolve_and_update_env_vars(self) -> "ConfigurableResourceFactory[TResValue]":
         """Processes the config dictionary to resolve any EnvVar values. This is called at runtime
         when the resource is initialized, so the user is only shown the error if they attempt to
         kick off a run relying on this resource.
@@ -481,7 +483,7 @@ class ConfigurableResource(
 
     def _resolve_and_update_nested_resources(
         self, context: InitResourceContext
-    ) -> "ConfigurableResource[TResValue]":
+    ) -> "ConfigurableResourceFactory[TResValue]":
         """Updates any nested resources with the resource values from the context.
         In this case, populating partially configured resources or
         resources that return plain Python types.
@@ -531,9 +533,35 @@ class ConfigurableResource(
         """Returns the object that this resource hands to user code, accessible by ops or assets
         through the context or resource parameters. This works like the function decorated
         with @resource when using function-based resources.
+        """
+        raise NotImplementedError()
 
-        Default behavior for new class-based resources is to return itself, passing
-        the actual resource object to user code.
+
+@experimental
+class ConfigurableResource(ConfigurableResourceFactory[TResValue]):
+    """Base class for Dagster resources that utilize structured config.
+
+    This class is a subclass of both :py:class:`ResourceDefinition` and :py:class:`Config`, and
+    provides a default implementation of the resource_fn that returns the resource itself.
+
+    Example:
+    .. code-block:: python
+
+        class WriterResource(ConfigurableResource):
+            prefix: str
+
+            def output(self, text: str) -> None:
+                print(f"{self.prefix}{text}")
+
+    """
+
+    def create_resource(self, context: InitResourceContext) -> TResValue:
+        """Returns the object that this resource hands to user code, accessible by ops or assets
+        through the context or resource parameters. This works like the function decorated
+        with @resource when using function-based resources.
+
+        For ConfigurableResource, this function will return itself, passing
+        the actual ConfigurableResource object to user code.
         """
         return cast(TResValue, self)
 
@@ -554,9 +582,11 @@ class PartialResource(
     Generic[TResValue], ResourceDefinition, AllowDelayedDependencies, MakeConfigCacheable
 ):
     data: Dict[str, Any]
-    resource_cls: Type[ConfigurableResource[TResValue]]
+    resource_cls: Type[ConfigurableResourceFactory[TResValue]]
 
-    def __init__(self, resource_cls: Type[ConfigurableResource[TResValue]], data: Dict[str, Any]):
+    def __init__(
+        self, resource_cls: Type[ConfigurableResourceFactory[TResValue]], data: Dict[str, Any]
+    ):
         resource_pointers, data_without_resources = separate_resource_params(data)
 
         MakeConfigCacheable.__init__(self, data=data, resource_cls=resource_cls)  # type: ignore  # extends BaseModel, takes kwargs
@@ -589,9 +619,14 @@ class PartialResource(
         return self._nested_resources
 
 
-ResourceOrPartial: TypeAlias = Union[ConfigurableResource[TResValue], PartialResource[TResValue]]
+ResourceOrPartial: TypeAlias = Union[
+    ConfigurableResourceFactory[TResValue], PartialResource[TResValue]
+]
 ResourceOrPartialOrValue: TypeAlias = Union[
-    ConfigurableResource[TResValue], PartialResource[TResValue], ResourceDefinition, TResValue
+    ConfigurableResourceFactory[TResValue],
+    PartialResource[TResValue],
+    ResourceDefinition,
+    TResValue,
 ]
 
 
@@ -603,7 +638,7 @@ class ResourceDependency(Generic[V]):
     def __set_name__(self, _owner, name):
         self._name = name
 
-    def __get__(self, obj: "ConfigurableResource", __owner: Any) -> V:
+    def __get__(self, obj: "ConfigurableResourceFactory", __owner: Any) -> V:
         return getattr(obj, self._name)
 
     def __set__(self, obj: Optional[object], value: ResourceOrPartialOrValue[V]) -> None:
@@ -652,7 +687,9 @@ class ConfigurableLegacyResourceAdapter(ConfigurableResource, ABC):
 
 
 @experimental
-class ConfigurableIOManagerFactory(ConfigurableResource[TIOManagerValue], IOManagerDefinition):
+class ConfigurableIOManagerFactory(
+    ConfigurableResourceFactory[TIOManagerValue], IOManagerDefinition
+):
     """Base class for Dagster IO managers that utilize structured config. This base class
     is useful for cases in which the returned IO manager is not the same as the class itself
     (e.g. when it is a wrapper around the actual IO manager implementation).
@@ -663,7 +700,7 @@ class ConfigurableIOManagerFactory(ConfigurableResource[TIOManagerValue], IOMana
     """
 
     def __init__(self, **data: Any):
-        ConfigurableResource.__init__(self, **data)
+        ConfigurableResourceFactory.__init__(self, **data)
         IOManagerDefinition.__init__(
             self,
             resource_fn=self.initialize_and_run,
@@ -688,7 +725,9 @@ class ConfigurableIOManagerFactory(ConfigurableResource[TIOManagerValue], IOMana
 
 
 class PartialIOManager(Generic[TResValue], PartialResource[TResValue], IOManagerDefinition):
-    def __init__(self, resource_cls: Type[ConfigurableResource[TResValue]], data: Dict[str, Any]):
+    def __init__(
+        self, resource_cls: Type[ConfigurableResourceFactory[TResValue]], data: Dict[str, Any]
+    ):
         PartialResource.__init__(self, resource_cls, data)
         IOManagerDefinition.__init__(
             self,
@@ -998,6 +1037,6 @@ def _call_resource_fn_with_default(obj: ResourceDefinition, context: InitResourc
 
 LateBoundTypesForResourceTypeChecking.set_actual_types_for_type_checking(
     resource_dep_type=ResourceDependency,
-    resource_type=ConfigurableResource,
+    resource_type=ConfigurableResourceFactory,
     partial_resource_type=PartialResource,
 )