+++ python_modules/dagster/dagster_tests/asset_defs_tests/test_external_asset_graph.py
@@ -104,12 +104,12 @@ def make_context(defs_attrs):
 def test_get_repository_handle():
     asset_graph = ExternalAssetGraph.from_workspace(make_context(["defs1", "defs2"]))
 
-    assert asset_graph.get_job_names(asset1.key) == ["__ASSET_JOB"]
+    assert asset_graph.get_materialization_job_names(asset1.key) == ["__ASSET_JOB"]
     repo_handle1 = asset_graph.get_repository_handle(asset1.key)
     assert repo_handle1.repository_name == "__repository__"
     assert repo_handle1.repository_python_origin.code_pointer.fn_name == "defs1"
 
-    assert asset_graph.get_job_names(asset1.key) == ["__ASSET_JOB"]
+    assert asset_graph.get_materialization_job_names(asset1.key) == ["__ASSET_JOB"]
     repo_handle2 = asset_graph.get_repository_handle(asset2.key)
     assert repo_handle2.repository_name == "__repository__"
     assert repo_handle2.repository_python_origin.code_pointer.fn_name == "defs2"
@@ -126,12 +126,14 @@ def test_cross_repo_dep_with_source_asset():
         ).repository_python_origin.code_pointer.fn_name
         == "defs1"
     )
+    assert asset_graph.get_materialization_job_names(AssetKey("asset1")) == ["__ASSET_JOB"]
     assert (
         asset_graph.get_repository_handle(
             AssetKey("downstream")
         ).repository_python_origin.code_pointer.fn_name
         == "downstream_defs"
     )
+    assert asset_graph.get_materialization_job_names(AssetKey("downstream")) == ["__ASSET_JOB"]
 
 
 def test_cross_repo_dep_no_source_asset():
@@ -147,12 +149,16 @@ def test_cross_repo_dep_no_source_asset():
         ).repository_python_origin.code_pointer.fn_name
         == "defs1"
     )
+    assert asset_graph.get_materialization_job_names(AssetKey("asset1")) == ["__ASSET_JOB"]
     assert (
         asset_graph.get_repository_handle(
             AssetKey("downstream_non_arg_dep")
         ).repository_python_origin.code_pointer.fn_name
         == "downstream_defs_no_source"
     )
+    assert asset_graph.get_materialization_job_names(AssetKey("downstream_non_arg_dep")) == [
+        "__ASSET_JOB"
+    ]
 
 
 def test_partitioned_source_asset():