+++ python_modules/dagster/dagster_tests/core_tests/serdes_tests/run_pdb.py
@@ -1,0 +1,5 @@
+import pdb
+
+foo = "HELLO"
+
+pdb.set_trace()